Conflicts:
	include/live/ultrasound.c

+++ include/live/servo.c
@@ -12,7 +12,7 @@ void servo_init(){
     pwm_enable();
 }
 
-int servo_set_pos(int angle){
+uint16_t servo_set_pos(int angle){
     if (angle >= 270){
         current_pos = 270;
         pwm_set_match(2, MAX_WIDTH);
@@ -27,7 +27,7 @@ int servo_set_pos(int angle){
     return current_pos;
 }
 
-int servo_get_pos(){
+uint16_t servo_get_pos(){
 #ifdef RECORD
 	network_send(SERVO_HEADER, &current_pos, sizeof(int), NULL);
 #endif
